Add shortcode maincontent
At the moment, using Ulysses to write the content. When exporting, need
to differentiate Hugo's index.md (contains frontmatter) with content.md
(which is exported from Ulysses)

The shortcode helps with the integration of content. We lose the ability
to have summary though, since the summary/toc is generated before the
shortcode is processed. It seems, few years back, it was the other way
around until someone created a PR. Should have made it configuration
instead.